// src/types/AnimeData.ts
import { Anime } from '@tutkli/jikan-ts';

export type IAnimeData = {
  data: Anime;
};

export interface ISearchData {
  pagination: {
    last_visible_page: number;
    has_next_page: boolean;
    current_page: number;
    items: {
      count: number;
      total: number;
      per_page: number;
    };
  };
  data: Anime[];
}